+++ frontend/src/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import Home from './Home';
-import About from './About';
+import Home from './Components/Home/Home';
+import About from './Components/About/About';
 import { BrowserRouter as Router } from 'react-router-dom';
 import registerServiceWorker from './registerServiceWorker';
 import { Route } from 'react-router-dom'
